convert arguments to strings check to see if they are equal
if they are return true, create two variables to hold empty arrays c, d, for loop and iterate over a, compare a[i] to b[i] if not equal, if not push into said variables we setup already, reverse d, if statement if c length ==2 and c==d return true. as we can only swap once hence c will be 2 unless we swapped more than that.